BluetoothTest and DatabaseTest removed.
