+++ mobile/src/main/AndroidManifest.xml
@@ -28,14 +28,10 @@
             </intent-filter>
         </activity>
         <activity android:name=".activities.devices.ShowDevicesActivity" />
-        <activity android:name=".activities.bluetoothTest.BluetoothTest" />
 
         <service
             android:name=".model.tasks.BluetoothConnection.BluetoothConnectionService"
             android:exported="false" />
-
-        <activity android:name=".activities.databaseTest.SqlActivity" />
-        <activity android:name=".activities.databaseTest.CouchebaseActivity"></activity>
     </application>
 
 </manifest>
