+++ mobile/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">WearNavigationSimple</string>
+    <string name="app_name">Wear Navigation</string>
     <string name="app_yes">Yes</string>
     <string name="app_no">No</string>
